Convert daemon into cargo workspace
Shares code between linux and windows variants.
