+++ .gitignore
@@ -3,7 +3,6 @@
 daemon/target/
 firmware/target/
 shared/target/
-win-service/target/
 
 # Remove Cargo.lock from gitignore if creating an executable, leave it for libraries
 # More information here https://doc.rust-lang.org/cargo/guide/cargo-toml-vs-cargo-lock.html
