+++ daemon/Cargo.toml
@@ -1,14 +1,7 @@
-[package]
-name = "hw-cpu-daemon"
-version = "0.1.0"
-authors = ["James Hillyerd <james@hillyerd.com>"]
-edition = "2018"
+[workspace]
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-postcard = { version = "0.5.1", features = ["alloc"] }
-shared = { path = "../shared" }
-systemstat = "0.1.6"
-serialport = "4.0.0"
-time = "0.2.25"
+members = [
+    "lib",
+    "linux",
+    "windows",
+]
