+++ daemon/lib/src/avg.rs
