+++ daemon/linux/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "linux"
+version = "0.1.0"
+authors = ["James Hillyerd <james@hillyerd.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+lib = { path = "../lib" }
